Add files via upload
Appendix B Summary statistics of the model communities that resulted in two or more guilds were compared across ecological–evolutionary scales in three environmental structures. In simulations, the environmental niche only affected the per capita birth–death rates. The model parameters were same as the scenarios presented in the main text, except for only the first population size of guilds was explored in the small system. (a) Functional uniqueness and redundancy were compared to the relative portions explained by environmental and pure spatial components in the total explained variation of overall model; then, the total explained variation and these relative portions were compared between overall and hierarchical models. The portions explained in hierarchical model were calculated for the scenarios in that the hierarchical guild structure was significant. vs.guild-habitat, the number of guilds coexisted in model communities was compared to the number of habitat types identified by hierarchical guild structure. (b) Power rate of all components in overall and hierarchical models. The number of model communities is provided within parentheses on the x-axis.   

Error in Appendix B submitted June 29, 2021: The number of model communities in the four humps environmental structure and dispersal rate m = 0.01 was 16 but not 17.