Add files via upload
Appendix C Constrained randomization of the PCNM variables with positive eigenvalues built from the environmental distance was applied to the scenarios presented in Figure 4. (a) Functional uniqueness and redundancy were compared to the relative portions explained by environmental and pure spatial components in total explained variation. (b) Type I error rate of environmental and pure environmental components and power rate of other components. The number of model communities is provided within parentheses on the x-axis.