Add files via upload
Appendix E Constrained randomization of the PCNM variables with positive eigenvalues built from the environmental distance and the hierarchical guild structure were applied to the scenarios presented in Appendix B. (a) Functional uniqueness and redundancy were compared to the relative portions explained by environmental and pure spatial components in the total explained variation of overall model; then, the total explained variation and these relative portions were compared between overall and hierarchical models. The portions explained in hierarchical model were calculated for the scenarios in that the hierarchical guild structure was significant. vs.guild-habitat, the number of guilds coexisted in model communities was compared to the number of habitat types identified by hierarchical guild structure. (b) Power rate of all components in overall and hierarchical models. The number of model communities is provided within parentheses on the x-axis.